// core/ec-core/umbracoClient/models/ImageColumnsPropertiesModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApiBlockGridModel } from './ApiBlockGridModel';
import type { RichTextModel } from './RichTextModel';
export type ImageColumnsPropertiesModel = {
    imageGrid?: ApiBlockGridModel;
    subtext?: RichTextModel;
};